tab1 get contacts done, needs improved UI
